Employer controller secured

+++ src/main/java/org/joe/jobpoertalapp/services/EmployerService.java
@@ -3,12 +3,16 @@ package org.joe.jobpoertalapp.services;
 import org.joe.jobpoertalapp.dtos.incoming.InJobDto;
 import org.joe.jobpoertalapp.dtos.outgoing.OutApplicationDto;
 import org.joe.jobpoertalapp.dtos.outgoing.OutJobDto;
+import org.joe.jobpoertalapp.entities.Application;
+import org.joe.jobpoertalapp.entities.Job;
 import org.joe.jobpoertalapp.entities.User;
+import org.joe.jobpoertalapp.exceptions.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class EmployerService {
@@ -19,11 +23,24 @@ public class EmployerService {
         this.applicationService = applicationService;
     }
 
-    public OutJobDto postJob(InJobDto inJobDto, Long employerId){
-        return jobService.postJob(inJobDto,employerId);
+    public OutJobDto postJob(InJobDto inJobDto){
+
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) auth.getPrincipal();
+        Long userId = user.getId();
+
+        return jobService.postJob(inJobDto,userId);
     }
 
     public void deleteJob(Long id){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) auth.getPrincipal();
+        Long userId = user.getId();
+
+        Job job = jobService.getJobById(id);
+        if(!Objects.equals(job.getEmployer().getId(), userId)){
+            throw new AccessDeniedException("Access denied");
+        }
         jobService.deleteJob(id);
     }
 
@@ -37,14 +54,50 @@ public class EmployerService {
     }
 
     public List<OutApplicationDto> getApplicationsByJobId(Long jobId){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
+        User user = (User) auth.getPrincipal();
+
+        Long userId = user.getId();
+
+        Job job = jobService.getJobById(jobId);
+        if (!Objects.equals(job.getEmployer().getId(), userId)){
+            throw new AccessDeniedException("Access Denied");
+        }
         return applicationService.viewApplicationByJobId(jobId);
     }
 
     public OutApplicationDto approveApplication(Long id){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
+        User user = (User) auth.getPrincipal();
+
+        Long userId = user.getId();
+
+        Application application = applicationService.getApplicationById(id);
+        Job job = jobService.getJobById(application.getJob().getId());
+
+        if (!Objects.equals(job.getEmployer().getId(), userId)){
+            throw new AccessDeniedException("Access Denied");
+        }
+
         return applicationService.approveApplication(id);
     }
 
     public OutApplicationDto rejectApplication(Long id){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
+        User user = (User) auth.getPrincipal();
+
+        Long userId = user.getId();
+
+        Application application = applicationService.getApplicationById(id);
+        Job job = jobService.getJobById(application.getJob().getId());
+
+        if (!Objects.equals(job.getEmployer().getId(), userId)){
+            throw new AccessDeniedException("Access Denied");
+        }
+
         return applicationService.rejectApplication(id);
     }
